Implemented cim xUnit tests 1. Added infrastructure for p/invoking into libmi.so 2. Renamed MMI.Native-stub.cs: removed -stub 3. Implemented basic unit tests
